update forked repo

+++ src/main/java/de/bonndan/nivio/output/map/svg/WobblyGroupOutline.java
@@ -24,7 +24,7 @@ public class WobblyGroupOutline {
         StringBuilder points = new StringBuilder("M");
         for (var i = 0; i < middles.size(); i++) {
 
-            var j = i+1;
+            var j = i + 1;
             if (i == middles.size() - 1) //when the last ist reached, reconnect to first
                 j = 0;
 
@@ -33,11 +33,11 @@ public class WobblyGroupOutline {
             var controlPoint = corners.get(j);
             var next = middles.get(j);
 
-            points.append(" ").append(prev.x).append(",").append(prev.y).append(" ");
-            points.append("Q ").append(controlPoint.x).append(",").append(controlPoint.y);
+            points.append(" ").append((int) prev.x).append(",").append((int) prev.y).append(" ");
+            points.append("Q ").append((int) controlPoint.x).append(",").append((int) controlPoint.y);
 
             if (j == 0) {
-                points.append(" ").append(next.x).append(",").append(next.y);
+                points.append(" ").append((int) next.x).append(",").append((int) next.y);
             }
 
         }
